[PATCH] arm64: crypto: Improve target optimization flags
The target flags which are needed to enable
extensions for cpu intructions were overriding
our global setting of "mcpu=cortex-a57" and targeting
a generic armv8-a cpus instead.

This patch removes these flags since we have
these extensions for cpu instructions enabled
globally.
By doing this, we cantarget "cortex-a57",
which is the closest cpu target to kryo,
while keeping the extensions for the CRC
and Crypto hardware instructions.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>