ASoC: wcd-mbhc: Remove ifdefs around detection improvement code
Commit feb2dac51759e46b2db6534e1da7667df8a2187e restricted
the improvements to the CONFIG_MACH_MSM8998_DUMPLING defconfig option
which is never true for us. I do now know the motivation behind
ifdef'ing the only supported device in a tree but that's not a
requirement here.

This is a separate patch and not squashed into the original
cherry-pick to keep it identical to the original commit from
Sultanxda's msm8998 kernel.

Signed-off-by: Harsh Shandilya <harsh@prjkt.io>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>