BACKPORT: sched/cpufreq: uclamp: add schedutil utilization clamping for RT and FAIR task
Currently schedutil enforces a maximum frequency when RT tasks are
RUNNABLE.  Such a mandatory policy can be made more tunable from
userspace thus allowing for example to define a max frequency which is
still reasonable for the execution of a specific RT workload.
This will contribute to make the RT class more friendly for power/energy
sensitive use-cases.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: TheHitMan7 <krtik.vrma@gmail.com>